long numbers as 16777216 wasn't calculated as expected cause the datatype wasn't suitable

+++ 0x08-recursion/5-sqrt_recursion.c
@@ -18,7 +18,7 @@ int _sqrt_recursion(int n)
 */
 int _sqrt(int n, int start, int end)
 {
-	int mid;
+	long mid;
 
 	if (n < 0)
 		return (-1);
